keep messing up pull requests and branch management

+++ config.py
@@ -1,5 +1,4 @@
 import torch
-
 use_cuda = False
 device = torch.device("cuda" if use_cuda else "cpu")
-dtype = torch.float32
+dtype = torch.float64
